correccion en el tipo de dato id en la clase Category

+++ IntegradorBackend/src/main/java/com/dh/grup8/IntegradorBackend/model/service/ICRUDService.java
@@ -3,9 +3,9 @@ package com.dh.grup8.IntegradorBackend.model.service;
 import java.util.Set;
 
 public interface ICRUDService<T> {
-    T findById(Integer id);
+    T findById(Long id);
     T create(T t);
-    void deleteById(Integer id);
+    void deleteById(Long id);
     T update(T t);
     Set<T> findAll();
 }
